# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "XChain Explorer API",
    "version": "1.0.0",
    "description": "Read-only REST API for the XChain Platform: tokens, balances, actions, markets, smart contracts, staking, attestations, and state checkpoints on Bitcoin, Litecoin, and Dogecoin.\n\nConventions: `coin` is the host chain, `tick` is a token symbol (there is no \"asset\" field). Amounts are arbitrary-precision decimal strings, and so are chain indices (`block_index`, `action_index`, `checkpoint_seq`, `snapshot_block`) on both the REST and WebSocket surfaces: they compare byte-for-byte across endpoints and stay exact above 2^53. List responses are `{total, data: [...]}` with `page`/`limit`/`sortorder` query parameters (limit max 100; balances/holders max 500). Errors are `{error: \"message\", code: \"STABLE_CODE\"}`. See the error-code registry at https://docs.xchain.io/protocol/Error_Codes.md. A WebSocket API lives at /{COIN}/api/websocket. See https://docs.xchain.io/components/explorer/WEBSOCKET.md.\n\nLLM-friendly docs: https://docs.xchain.io/llms.txt",
    "license": {
      "name": "AGPL-3.0-or-later",
      "url": "https://docs.xchain.io/legal/LICENSING.md"
    }
  },
  "servers": [
    {
      "url": "https://explorer.xchain.io"
    }
  ],
...
  "paths": {
...
    "/{COIN}/api/network": {
      "get": {
        "operationId": "network",
        "tags": [
          "Network"
        ],
        "summary": "Network statistics",
        "parameters": [
          {
            "$ref": "#/components/parameters/coin"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/CoinUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "coin": {
        "name": "COIN",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "enum": [
            "BTC",
            "TBTC",
            "RBTC",
            "LTC",
            "TLTC",
            "RLTC",
            "DOGE",
            "TDOGE",
            "RDOGE"
          ]
        },
        "description": "Chain + network: BTC/LTC/DOGE mainnet; T prefix = testnet, R prefix = regtest"
      },
...
    },
    "schemas": {
...
      "NetworkAsset": {
        "type": "object",
        "description": "A coin or token identity with its quoted prices (decimal strings).",
        "properties": {
          "name": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "price": {
            "type": "object",
            "properties": {
              "btc": {
                "type": "string",
                "description": "Price in BTC, 8dp decimal string"
              },
              "usd": {
                "type": "string",
                "description": "Price in USD, decimal string (\"0.00\" when no market)"
              }
            },
            "required": [
              "btc",
              "usd"
            ]
          }
        },
        "required": [
          "name",
          "symbol",
          "price"
        ]
      },
      "NetworkResponse": {
        "type": "object",
        "description": "Network statistics for one coin: indexer tip, mempool depth, per-action-type record counts, fee tiers, coin/XCHAIN identity and prices, and recommended finality depths.",
        "properties": {
          "network": {
            "type": "object",
            "description": "This coin's indexer tip. `block` is the height the dashboard checkpoint-stall alert reads.",
            "properties": {
              "block": {
                "type": "integer",
                "description": "Indexer tip height"
              },
              "time": {
                "type": "integer",
                "description": "Unix time of the tip block (0 when the chain is empty)"
              },
              "unconfirmed": {
                "type": "integer",
                "description": "Mempool depth from the decoder DB (0 when unreachable)"
              }
            },
            "required": [
              "block",
              "time",
              "unconfirmed"
            ]
          },
          "totals": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Exact record count per action table, keyed by table name"
          },
          "fee": {
            "type": "object",
            "description": "Suggested fee tiers in sat/vByte from this coin's encoder",
            "properties": {
              "low": {
                "type": "number"
              },
              "medium": {
                "type": "number"
              },
              "high": {
                "type": "number"
              }
            },
            "required": [
              "low",
              "medium",
              "high"
            ]
          },
          "coin": {
            "$ref": "#/components/schemas/NetworkAsset"
          },
          "xchain": {
            "$ref": "#/components/schemas/NetworkAsset"
          },
          "finality": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Recommended confirmation depth per chain (display guidance, never a gate)"
          },
          "runtime": {
            "type": "string",
            "description": "Server-side render time, present on every JSON response"
          }
        },
        "required": [
          "network",
          "totals",
          "fee",
          "coin",
          "xchain",
          "finality"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable message"
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable code (see https://docs.xchain.io/protocol/Error_Codes.md)"
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "CoinUnavailable": {
        "description": "Coin not configured on this explorer",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}